+++ data/src/main/java/com/choimaro/data/local/LocalDataSource.kt
@@ -1,10 +1,27 @@
 package com.choimaro.data.local
 
+import com.choimaro.data.db.dao.BookMarkDAO
 import com.choimaro.data.db.entity.ImageBookMarkEntity
+import javax.inject.Inject
 
-interface LocalDataSource {
-    suspend fun getAllBookMark(): List<ImageBookMarkEntity>
-    suspend fun insertBookMark(imageBookMarkEntity: ImageBookMarkEntity): Boolean
-    suspend fun deleteBookMark(ids: List<String>): Boolean
-    suspend fun deleteAll()
+class LocalDataSource @Inject constructor(
+    private val bookMarkDAO: BookMarkDAO
+) {
+    suspend fun getAllBookMark(): List<ImageBookMarkEntity> {
+        return bookMarkDAO.getAllBookMark()
+    }
+
+    suspend fun insertBookMark(imageBookMarkEntity: ImageBookMarkEntity): Boolean {
+        val insertCount = bookMarkDAO.insert(imageBookMarkEntity)
+        return insertCount > 0
+    }
+
+    suspend fun deleteBookMark(ids: List<String>): Boolean {
+        val deletedCount = bookMarkDAO.delete(ids)
+        return deletedCount == ids.size
+    }
+
+    suspend fun deleteAll() {
+        return bookMarkDAO.deleteAll()
+    }
 }
